Added AlarmCommand and its parser; Updated tests and models

+++ src/test/java/seedu/address/logic/commands/PurchaseMedicineViaPathCommandTest.java
@@ -9,6 +9,8 @@ import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 
+import java.math.BigDecimal;
+
 public class PurchaseMedicineViaPathCommandTest {
 
     private Model model;
@@ -26,7 +28,8 @@ public class PurchaseMedicineViaPathCommandTest {
     public void purchaseValidMedicine() {
         try {
             CommandResult commandResult =
-                    new PurchaseMedicineViaPathCommand(new String[] {"root", "TCM", "Panaddol"}, 50, 500)
+                    new PurchaseMedicineViaPathCommand(
+                            new String[] {"root", "TCM", "Panaddol"}, 50, BigDecimal.valueOf(500))
                     .execute(model, commandHistory);
             Assert.assertEquals("Purchase successful.", commandResult.getFeedbackToUser());
         } catch (CommandException ex) {
@@ -38,7 +41,8 @@ public class PurchaseMedicineViaPathCommandTest {
     public void invalidMedicinePath_throwCommandException() {
         try {
             CommandResult commandResult =
-                    new PurchaseMedicineViaPathCommand(new String[] {"root", "Panaddol"}, 50, 500)
+                    new PurchaseMedicineViaPathCommand(
+                            new String[] {"root", "Panaddol"}, 50, BigDecimal.valueOf(500))
                     .execute(model, commandHistory);
             Assert.fail();
         } catch (CommandException ex) {
@@ -50,7 +54,8 @@ public class PurchaseMedicineViaPathCommandTest {
     public void invalidChangeAmount_throwCommandException() {
         try {
             CommandResult commandResult =
-                    new PurchaseMedicineViaPathCommand(new String[] {"root", "TCM", "Panaddol"}, -50, 500)
+                    new PurchaseMedicineViaPathCommand(
+                            new String[] {"root", "TCM", "Panaddol"}, -50, BigDecimal.valueOf(500))
                     .execute(model, commandHistory);
             Assert.fail();
         } catch (CommandException ex) {
